requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Elasticsearch
elasticsearch==8.11.0

# База данных MySQL для загрузки данных FИАС
mysql-connector-python==8.2.0

# Утилиты
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7

# Обработка текста
regex==2023.10.3
unidecode==1.3.7
